requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.1.1
idna==3.10
msgpack==1.1.0
py==1.11.0
pydig==0.4.0
pymetasploit3==1.0.6
python-nmap==0.7.1
python-owasp-zap-v2.4==0.1.0
python3-nmap==1.9.1
requests==2.32.3
retry==0.9.2
simplejson==3.19.3
six==1.17.0
urllib3==2.3.0
zaproxy==0.4.0
